refactor: namings to increase readability, split isMeteorComponent into new utility file
