+++ src/Administration/Resources/app/administration/src/core/service/utils/field-inheritance.utils.ts
@@ -21,7 +21,7 @@ type FieldInheritanceDefinition = {
  * @sw-package framework
  * @private
  */
-export const mapInheritanceComponentNames = [
+export const componentNamesSupportingMapInheritance = [
     'sw-checkbox-field',
     'sw-colorpicker',
     'sw-compact-colorpicker',
@@ -44,40 +44,11 @@ export const mapInheritanceComponentNames = [
  * @sw-package framework
  * @private
  */
-export const fieldHandlingInheritanceItselfComponentNames = [
+export const componentNamesHandlingInheritanceThemselves = [
     'mt-switch',
     'mt-checkbox',
     'sw-switch-field',
     'sw-checkbox-field',
-];
-
-/**
- * @sw-package framework
- * @private
- */
-export const meteorComponentNames = [
-    'mt-checkbox',
-    'mt-colorpicker',
-    'mt-datepicker',
-    'mt-email-field',
-    'mt-number-field',
-    'mt-password-field',
-    'mt-select',
-    'mt-switch',
-    'mt-text-field',
-    'mt-textarea',
-    'mt-url-field',
-    // internally these map to their meteor component equivalent
-    'sw-checkbox-field',
-    'sw-colorpicker',
-    'sw-datepicker',
-    'sw-number-field',
-    'sw-password-field',
-    'sw-switch-field',
-    'sw-text-editor',
-    'sw-text-field',
-    'sw-textarea-field',
-    'sw-url-field',
 ];
 
 function isSupported(field: FieldInheritanceDefinition | null, componentNames: string[]) {
@@ -91,15 +62,7 @@ function isSupported(field: FieldInheritanceDefinition | null, componentNames: s
  * @private
  */
 export function supportsMapInheritance(field: FieldInheritanceDefinition | null) {
-    return isSupported(field, mapInheritanceComponentNames);
-}
-
-/**
- * @sw-package framework
- * @private
- */
-export function isMeteorComponent(field: FieldInheritanceDefinition | null) {
-    return isSupported(field, meteorComponentNames);
+    return isSupported(field, componentNamesSupportingMapInheritance);
 }
 
 /**
@@ -107,5 +70,5 @@ export function isMeteorComponent(field: FieldInheritanceDefinition | null) {
  * @private
  */
 export function isFieldHandlingInheritanceItself(field: FieldInheritanceDefinition | null) {
-    return isSupported(field, fieldHandlingInheritanceItselfComponentNames);
+    return isSupported(field, componentNamesHandlingInheritanceThemselves);
 }
