+++ src/Administration/Resources/app/administration/src/module/sw-settings/component/sw-system-config/index.js
@@ -3,10 +3,8 @@
  */
 import ErrorResolverSystemConfig from 'src/core/data/error-resolver.system-config.data';
 import { deepCloneWithEntity } from 'src/core/service/extension-api-data.service';
-import {
-    isMeteorComponent as isMeteorFieldComponent,
-    supportsMapInheritance,
-} from 'src/core/service/utils/field-inheritance.utils';
+import { supportsMapInheritance } from 'src/core/service/utils/field-inheritance.utils';
+import { isMeteorComponent as isMeteorFieldComponent } from 'src/core/service/utils/meteor-component.utils';
 import { isFieldHandlingLabelAndHelpText as fieldHandlesLabelAndHelpText } from 'src/core/service/utils/field-label.utils';
 import template from './sw-system-config.html.twig';
 import './sw-system-config.scss';
@@ -308,7 +306,7 @@ export default {
         },
 
         isFieldHandlingLabelAndHelpText(element) {
-            return fieldHandlesLabelAndHelpText(element, { resolveType: true });
+            return fieldHandlesLabelAndHelpText(element, { renderedByFormFieldRenderer: true });
         },
 
         getMeteorElementBind(element, mapInheritance) {
